Enable Vagrant support (#29)
* Improve python package installation

pip-tools allows to track dependencies on python packages, helping on the
upgrade process. This change intruduces the pip-tools usage as well as improves
the method to consume the python binaries through the pip installation.

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Collect galaxy requirements in a file

ansible-galaxy allows to install roles and collections listed in a
requirements file. This change uses that functionality as well as defines
the versions for every collection used at this point.

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Refactor gce_run.sh script

This refactor tasks allows to manage all the working files on the home directory
and reduces some repetitive instructions.

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Change Ubuntu Release to Focal

The gtp5g module used by free5GC project requires a kernel version <5.4

https://github.com/free5gc/gtp5g/blob/master/README.md

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Add minimal config file for one-summit-22-workshop

The nephio-ansible-install playbooks require a configuration file to
operate. This change intruduces a minimal sample file for its operation

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Enable Vagrant support

Vagrant tool allows the creation of local enviroment through the VMs. This
automation reduces the development time and accelates the development. This
changes introduces the Vagrantfile to enable that support.

Signed-off-by: Victor Morales <v.morales@samsung.com>

* Enable shell formatter

The shfmt tool helps to provide a format to bash scripts. This standarization
helps to detect some issues. This change facilitates the formatting process.

Signed-off-by: Victor Morales <v.morales@samsung.com>

---------

Signed-off-by: Victor Morales <v.morales@samsung.com>